ml_service/requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
joblib
transformers
pillow
torch
timm
deep_translator